refactor(docs-retrieval): extract package-content + resolver singleton (final)
Final PR of the content-fetcher modularization stack. Moves the package-content
integration — including the module-level `_packageResolver` singleton — out of
content-fetcher.ts into `content-fetcher/package-content.ts`. content-fetcher.ts
is now a ~290-line orchestrator: `fetchContent` + `wrapContentAsJsonGuide` +
the small URL/content-type helpers, composing the extracted loaders and
transport.

## What moved (verbatim, no logic change)

- `_packageResolver` singleton + `setPackageResolver`
- `derivePathSlug`, `buildMilestoneWebsiteUrl`
- `resolvePackageMilestones`, `resolvePackageNavLinks`
- `isPathManifest`, `getManifestMilestoneIds`
- `fetchPackageContent` (incl. the `Promise.all` interleaving of content fetch /
  milestone resolution / baseUrl resolution — preserved exactly), `fetchPackageById`

## Breaking the cycle (reviewer gotcha)

`package-content.ts` imports `fetchContent` from `../content-fetcher`, and
`fetchPackageContent` calls `injectJourneyExtrasIntoJsonGuide` (imported from
`./cover-page`). The dependency is **one-directional**: content-fetcher.ts no
longer references any package code, so there is no import cycle. To keep that
true, the barrel's package re-export block now pulls these names from
`./content-fetcher/package-content` instead of `./content-fetcher` — the
**public surface is unchanged** (same six names exported from
`docs-retrieval`), only the internal source path moved.

content-fetcher.ts sheds five now-unused imports (`PackageResolver`,
`ResolvedNavLink`, `getPackageRenderType`, `Milestone`, `LearningJourneyMetadata`).

## Tests

The existing `package-content.test.ts` (36 tests) is the invariant safety net —
its assertions are untouched. Two moved-seam fixes only: the package-fn import
repoints to `./content-fetcher/package-content` (with `fetchContent` still from
`./content-fetcher`). Added happy-path coverage for the two previously-untested
seams: `resolvePackageNavLinks` (resolution, title fallback, skip-unresolvable,
empty/no-resolver guards) and `buildMilestoneWebsiteUrl` via
`resolvePackageMilestones` with a path slug (prefix match / no match / no slug).

A backup branch `refactor-content-fetcher-pre-pr9` was cut before this change.
